+++ src/UIComponents/atom/ConfirmBox/ConfirmBox.js
@@ -3,7 +3,14 @@ import PropTypes from "prop-types";
 import Button from "../Button";
 import "./ConfirmBox.scss";
 
-const ConfirmBox = ({ children, title, onConfirm, onCancel, position }) => {
+const ConfirmBox = ({
+  children,
+  title,
+  onConfirm,
+  onCancel,
+  position,
+  ...others
+}) => {
   const [showConfirmBox, setShowConfirmBox] = useState(false);
 
   const handleOk = () => {
@@ -16,16 +23,18 @@ const ConfirmBox = ({ children, title, onConfirm, onCancel, position }) => {
     if (onCancel) onCancel();
   };
 
-  const popoverPosition = position === "center" ? "0px" : "calc(100% + 4px)";
+  const customStyles =
+    position === "center"
+      ? { right: "0px" }
+      : position === "left"
+      ? { right: "calc(100% + 4px)" }
+      : { left: "calc(100% + 4px)" };
 
   return (
     <span className="relative">
       <span onClick={() => setShowConfirmBox(true)}>{children}</span>
       {showConfirmBox && (
-        <div
-          className="confirm-box-container"
-          style={{ right: popoverPosition }}
-        >
+        <div {...others} className="confirm-box-container" style={customStyles}>
           <span className="confirm-box-title">{title}</span>
           <span className="confirm-box-actions">
             <Button size="sm" onClick={handleOk}>
@@ -48,7 +57,7 @@ ConfirmBox.defaultProps = {
 
 ConfirmBox.propTypes = {
   title: PropTypes.string,
-  position: PropTypes.string,
+  position: PropTypes.oneOf(["center", "left", "right"]),
 };
 
 export default ConfirmBox;